pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comms-ai-datagen"
version = "1.0.0"
description = "Synthetic data generator for Central Bank Communications AI (Federal Reserve / FRBSF)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.34.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "click>=8.1",
]

[project.scripts]
datagen = "datagen.main:cli"

[tool.setuptools.packages.find]
where = ["src"]
